normalized_ranges: Revert Sequence check
This breaks if the user passes a tuple of ranges (which is pretty normal for single RFS calls).

The reason this breaks is because `Tuple`s and `List`s both inherit from `Sequence`, so either is an instance of `Sequence`. Blame mypy for flagging this.

PS: setsu pls2write basic tests