Improve Transfer heuristics (#79)
SMPTE2084 is not a safe default, as this implies HDR. An HDR source
should already have SMPTE2084 along with other HDR metadata set, or else
it will look completely washed out. This means that if we have a video
with no transfer characteristics set, it is much safer to assume that it
is SDR.

Although BT709/1886 and both BT2020 transfers are functionally
identical, this change sets the proper BT2020 transfer based on bit
depth in order to be technically correct.