depth: Change auto-dithering to DitherType.ORDERED
Courtesy of @wiwaz. From our PMs:

@wiwaz:
```
i did tests
at high bitdepths the difference between ordered and error diffusion spatially is unnoticeable
but ordered is static
using error diffusion all the time is hurting our compressability
and maybe even making quality worse
since its adding random temporal noise that we really dont even need
again, the difference is unnoticeable spatially at high depths
ordered only looks ugly when dithered down to 8
so basically when we are doing 32 to 16 or 16 to 10 whatever we should really be considering ordered instead
and opting for error diffusion only for the final dither down
```

With this in mind, this PR is more of a call to action for additional testers to see how well this works out in real-life application. Please test! @NSQY @Setsugennoao