Fix crash when using vskernels.LinearLight class on HDR video
When instantiating the LinearLight class on a video using the PQ transfer function, the error "ValueError: 16 is not a valid Matrix" is currently thrown. This is due to the transfer-to-matrix mappings not being filled in, and currently only working correctly if the transfer and matrix int values are identical. Filling in this map resolves the crash.