Reenable mypy step
Ignoring the errors for now because there are too many of them. Added
a problem matcher to bubble them up to GitHub.

Also added a basic smoke test, since a larger test suite was rejected,
to ensure that the package can be imported at all (catching things like
circular imports).