ChromaLocation.from_res: Always return LEFT
The decision to always return LEFT is simple:

Every regular format expects LEFT. The only exception is UHD, but only if that has HDR. SDR streams are explicitly tagged as being LEFT in rips we've checked.

As we've decided with Matrix.from_res, the onus should be on the user to ensure their UHD sources are properly tagged.