Small fix for type checking with pyright
The type checker that is included with Visual Studio Code (pyright)
will only resolve imports in the try branch of a try-except statement.

This means that VSPresetVideoFormat won't be resolved if your typing
stubs are up-to-date, as the correct import is in the except branch.

Swapping the try and except branches makes it work with freshly
generated stubs, but will break older ones. However, since there is
(seemingly) no way to make it work with both, having newer ones work
seems like the better compromise to me.